+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
+## Version 0.2.3x
+
+- Added string representation of arrays
+- Improvements in array comparisons (assertEquals())
+- Several bugfixes
+
 ## Version 0.2.3
 
 - Added captureException() to DSL to check for expected exceptions
 - Configurable checkstyle penalties (style_penalties.json)
-- Added string representation of arrays
 - Several bugfixes
 
 ## Version 0.2.2